added LED test sysex command

+++ firmware/src/application/src/devctl/devctl.h
@@ -2,6 +2,6 @@
 
 #include <stdint.h>
 
-int  DEVCTL_isDeviceControlMsg(uint8_t** const pBuff, uint32_t* const pLen);
-void DEVCTL_init(void);
-void DEVCTL_processMsg(uint8_t* buff, uint32_t len);
+uint16_t DEVCTL_isDeviceControlMsg(uint8_t** const pBuff, uint32_t* const pLen);
+void     DEVCTL_init(void);
+void     DEVCTL_processMsg(uint8_t* buff, uint32_t len);
